refactor: move files around
move test of specific "plugin" into their own file
