+++ src/function/getSpyFunction.ts
