+++ src/function/getStubFunction.ts
