+++ src/spec.spec.ts
@@ -7,9 +7,7 @@ import {
   simpleCallback,
   fetch,
   literalCallback,
-  promise,
   synchronous,
-  childProcess,
   delayed
 } from './specTestSuites'
 
@@ -200,7 +198,6 @@ test('replay on not saved input will spy', async t => {
 
 //#endregion
 
-
 //#region fetch
 test('fetch verify', async t => {
   const speced = await spec(fetch.success)
@@ -356,89 +353,6 @@ test('literalCallback fail case replay', async t => {
 })
 //#endregion
 
-//#region promise
-test('promise verify', async t => {
-  const speced = await spec(promise.success)
-  // not using `await` to make sure the return value is a promise.
-  // `await` will hide the error if the return value is not a promise.
-  return promise.increment(speced.subject, 2)
-    .then(actual => {
-      t.is(actual, 3)
-      return speced.satisfy([
-        { type: 'invoke', payload: ['increment', 2] },
-        { type: 'return', payload: {}, meta: { type: 'promise' } },
-        { type: 'promise', payload: 3, meta: { type: 'resolve' } }
-      ])
-    })
-})
-
-test('promise verify save', async t => {
-  const speced = await spec(promise.success, { id: 'promise', mode: 'save' })
-  return promise.increment(speced.subject, 2)
-    .then(actual => {
-      t.is(actual, 3)
-      return speced.satisfy([
-        { type: 'invoke', payload: ['increment', 2] },
-        { type: 'return', payload: {}, meta: { type: 'promise' } },
-        { type: 'promise', payload: 3, meta: { type: 'resolve' } }
-      ])
-    })
-})
-
-test('promise verify replay', async t => {
-  const speced = await spec(promise.success, { id: 'promise', mode: 'replay' })
-  return promise.increment(speced.subject, 2)
-    .then(actual => {
-      t.is(actual, 3)
-      return speced.satisfy([
-        { type: 'invoke', payload: ['increment', 2] },
-        { type: 'return', payload: {}, meta: { type: 'promise' } },
-        { type: 'promise', payload: 3, meta: { type: 'resolve' } }
-      ])
-    })
-})
-
-test('promise rejected verify', async t => {
-  const speced = await spec(promise.fail)
-  return promise.increment(speced.subject, 2)
-    .then(() => t.fail())
-    .catch(() => {
-      return speced.satisfy([
-        { type: 'invoke', payload: ['increment', 2] },
-        { type: 'return', payload: {}, meta: { type: 'promise' } },
-        { type: 'promise', payload: { message: 'fail' }, meta: { type: 'reject' } }
-      ])
-    })
-})
-
-test('promise rejected save', async t => {
-  const speced = await spec(promise.fail, { id: 'promise fail', mode: 'save' })
-  return promise.increment(speced.subject, 2)
-    .then(() => t.fail())
-    .catch(() => {
-      return speced.satisfy([
-        { type: 'invoke', payload: ['increment', 2] },
-        { type: 'return', payload: {}, meta: { type: 'promise' } },
-        { type: 'promise', payload: { message: 'fail' }, meta: { type: 'reject' } }
-      ])
-    })
-})
-
-test('promise rejected replay', async t => {
-  const speced = await spec(promise.fail, { id: 'promise fail', mode: 'replay' })
-  return promise.increment(speced.subject, 2)
-    .then(() => t.fail())
-    .catch(() => {
-      return speced.satisfy([
-        { type: 'invoke', payload: ['increment', 2] },
-        { type: 'return', payload: {}, meta: { type: 'promise' } },
-        { type: 'promise', payload: { message: 'fail' }, meta: { type: 'reject' } }
-      ])
-    })
-})
-//#endregion
-
-
 //#region synchronous
 test('synchronous verify', async t => {
   const speced = await spec(synchronous.success)
@@ -505,111 +419,4 @@ test('synchronous fail replay', async t => {
   ])
 })
 
-//#endregion
-
-//#region childProcess
-test('childProcess verify', async t => {
-  const speced = await spec(childProcess.spawnSuccess)
-  const actual = await childProcess.increment(speced.subject, 2)
-  t.deepEqual(actual, {
-    result: [['stdout', 3], ['stdout', 4], ['stdout', 5]],
-    code: 0
-  })
-
-  await speced.satisfy([
-    { type: 'invoke', payload: ['increment', [2]] },
-    { type: 'return', payload: {}, meta: { type: 'childProcess' } },
-    { type: 'childProcess', payload: [3], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [4], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [5], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [0], meta: { site: ['on'], event: 'close' } }
-  ])
-})
-
-test('childProcess save', async t => {
-  const speced = await spec(childProcess.spawnSuccess, { id: 'childProcess/success', mode: 'save' })
-  const actual = await childProcess.increment(speced.subject, 2)
-  t.deepEqual(actual, {
-    result: [['stdout', 3], ['stdout', 4], ['stdout', 5]],
-    code: 0
-  })
-
-  await speced.satisfy([
-    { type: 'invoke', payload: ['increment', [2]] },
-    { type: 'return', payload: {}, meta: { type: 'childProcess' } },
-    { type: 'childProcess', payload: [3], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [4], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [5], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [0], meta: { site: ['on'], event: 'close' } }
-  ])
-})
-
-test('childProcess replay', async t => {
-  const speced = await spec(childProcess.spawnSuccess, { id: 'childProcess/success', mode: 'replay' })
-  const actual = await childProcess.increment(speced.subject, 2)
-  t.deepEqual(actual, {
-    result: [['stdout', 3], ['stdout', 4], ['stdout', 5]],
-    code: 0
-  })
-
-  await speced.satisfy([
-    { type: 'invoke', payload: ['increment', [2]] },
-    { type: 'return', payload: {}, meta: { type: 'childProcess' } },
-    { type: 'childProcess', payload: [3], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [4], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [5], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [0], meta: { site: ['on'], event: 'close' } }
-  ])
-})
-
-test('childProcess fail case verify', async t => {
-  const speced = await spec(childProcess.spawnFail)
-  const actual = await childProcess.increment(speced.subject, 2)
-  t.deepEqual(actual, {
-    result: [['stdout', 3], ['stderr', 4]],
-    code: 1
-  })
-
-  await speced.satisfy([
-    { type: 'invoke', payload: ['increment', [2]] },
-    { type: 'return', payload: {}, meta: { type: 'childProcess' } },
-    { type: 'childProcess', payload: [3], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [4], meta: { site: ['stderr', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [1], meta: { site: ['on'], event: 'close' } }
-  ])
-})
-
-test('childProcess fail case save', async t => {
-  const speced = await spec(childProcess.spawnFail, { id: 'childProcess/fail', mode: 'save' })
-  const actual = await childProcess.increment(speced.subject, 2)
-  t.deepEqual(actual, {
-    result: [['stdout', 3], ['stderr', 4]],
-    code: 1
-  })
-
-  await speced.satisfy([
-    { type: 'invoke', payload: ['increment', [2]] },
-    { type: 'return', payload: {}, meta: { type: 'childProcess' } },
-    { type: 'childProcess', payload: [3], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [4], meta: { site: ['stderr', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [1], meta: { site: ['on'], event: 'close' } }
-  ])
-})
-
-test('childProcess fail case replay', async t => {
-  const speced = await spec(childProcess.spawnFail, { id: 'childProcess/fail', mode: 'replay' })
-  const actual = await childProcess.increment(speced.subject, 2)
-  t.deepEqual(actual, {
-    result: [['stdout', 3], ['stderr', 4]],
-    code: 1
-  })
-
-  await speced.satisfy([
-    { type: 'invoke', payload: ['increment', [2]] },
-    { type: 'return', payload: {}, meta: { type: 'childProcess' } },
-    { type: 'childProcess', payload: [3], meta: { site: ['stdout', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [4], meta: { site: ['stderr', 'on'], event: 'data' } },
-    { type: 'childProcess', payload: [1], meta: { site: ['on'], event: 'close' } }
-  ])
-})
 //#endregion
